requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
cryptography==41.0.7
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
pandas==2.0.3
requests==2.31.0